Refacter CT log client to use interface
